requirements.txt
pandas
plotly-express
streamlit
ipython
nbformat
